deployment/dashboard/requirements.txt
flask==2.0.1
pandas==1.0.5
nltk==3.5
h5py==2.10.0
requests
google-cloud-bigquery
